Improve the image build for speed
Splits integration tests into an earlier stage of the build process.
Unit tests and linting are performed in a later stage to ensure that
only code that compiles is merged to master.

This means that when running `npm run test:integration` only integration
tests will be built, increasing development test speed.