Allow multiple testfs instances to be set at the same time
In previous versions, only one testfs instance was allowed to be
set at the time. This was enforced by a global lock to prevent causing
filesystem inconsistencies if restoring the same set of files.

While this technically worked, it was too restrictive, as users may have
wanted to have a global testfs instance for all tests, while having
local testfs instances for specific tests.

With this new feature, multiple testfs instances are allowed as long as
they are restored in reverse order as they were set.

Change-type: minor